+++ tests/TestCase/Core/PluginTest.php
@@ -28,20 +28,20 @@ class PluginTest extends TestCase
     {
         $result = Plugin::all();
         $this->assertEquals(ME_CMS, $result[0]);
-        $this->assertEquals(ME_TOOLS, $result[1]);
+        $this->assertEquals('MeTools', $result[1]);
         $this->assertNotContains('TestPlugin', $result);
 
         Plugin::load('TestPlugin');
         $result = Plugin::all();
         $this->assertEquals(ME_CMS, $result[0]);
-        $this->assertEquals(ME_TOOLS, $result[1]);
+        $this->assertEquals('MeTools', $result[1]);
         $this->assertContains('TestPlugin', $result);
 
         $result = Plugin::all(['order' => false]);
         $this->assertNotEquals(ME_CMS, $result[0]);
-        $this->assertNotEquals(ME_TOOLS, $result[1]);
+        $this->assertNotEquals('MeTools', $result[1]);
         $this->assertContains(ME_CMS, $result);
-        $this->assertContains(ME_TOOLS, $result);
+        $this->assertContains('MeTools', $result);
         $this->assertContains('TestPlugin', $result);
     }
 }