+++ tests/TestCase/View/View/AppViewTest.php
@@ -91,16 +91,16 @@ class AppViewTest extends TestCase
         }, $this->View->helpers()->loaded());
 
         $this->assertEquals([
-            ME_TOOLS . '\View\Helper\HtmlHelper',
-            ME_TOOLS . '\View\Helper\DropdownHelper',
-            ME_TOOLS . '\View\Helper\FormHelper',
-            ME_TOOLS . '\View\Helper\LibraryHelper',
-            ME_TOOLS . '\View\Helper\PaginatorHelper',
-            ASSETS . '\View\Helper\AssetHelper',
-            THUMBER . '\View\Helper\ThumbHelper',
+            'MeTools\View\Helper\HtmlHelper',
+            'MeTools\View\Helper\DropdownHelper',
+            'MeTools\View\Helper\FormHelper',
+            'MeTools\View\Helper\LibraryHelper',
+            'MeTools\View\Helper\PaginatorHelper',
+            'Assets\View\Helper\AssetHelper',
+            'Thumber\View\Helper\ThumbHelper',
             'WyriHaximus\MinifyHtml\View\Helper\MinifyHtmlHelper',
-            ME_TOOLS . '\View\Helper\BBCodeHelper',
-            ME_TOOLS . '\View\Helper\BreadcrumbsHelper',
+            'MeTools\View\Helper\BBCodeHelper',
+            'MeTools\View\Helper\BreadcrumbsHelper',
             RECAPTCHA_MAILHIDE . '\View\Helper\MailhideHelper',
             ME_CMS . '\View\Helper\WidgetHelper',
         ], $helpers);