+++ tests/test_app/TestApp/config/routes.php
@@ -3,6 +3,6 @@ use Cake\Routing\Router;
 
 Router::scope('/', function ($routes) {
     $routes->loadPlugin(RECAPTCHA_MAILHIDE);
-    $routes->loadPlugin(THUMBER);
+    $routes->loadPlugin('Thumber');
     $routes->loadPlugin(ME_CMS);
 });